hwaddrs: Rip out use of rand()
rand() has a non-zero chance of getting you something slightly
random-ish, but there is a good chance of getting the same thing everyone
else gets.  In particular since the clock is used as a seed `hwaddrs` is
run early enough in system start for the clock to be hazardously
predictable.

Since they shouldn't reproduce, securely hash the build fingerprint and
device serial number together to produce a reasonable substitute.

Change-Id: I12ff2baaaae313f57823264e0ec6e621aa468f05