Update patch set 2
Patch Set 2:

I don't know if all LG devices have a /persist partition and if they all store their mac addresses in /persist. The only device I own is a h850 and I have access to a friend who owns a us996.

The /persist/.macaddr and /persist/.baddr files do not have anything to do with camera and sensors, so the patch description isn't a good description.

Sorry if this might sound like a dumb question but what is the purpose of this patch?

For the msm8996 devices I think this patch would require changes of the device dependent selinux policy (which should be doable). Is this change part of a larger reorganization such as moving the hwaddr binary to the LG common repo?

If this helps to prevent the ugly workarounds related to the misc partition access in https://review.lineageos.org/c/LineageOS/android_device_lge_msm8996-common/+/266190 then this might be a good thing.

Patch-set: 2
Reviewer: Gerrit User 25069 <25069@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>