Update patch set 2
Patch Set 2:

> Patch Set 2: Verified-1 Code-Review-2
> 
> Not really a good idea. This change is already in the msm8996 tree. Other devices that need it will do their own. I'm not even sure the hwaddrs is even applicable for all the LGE devices even. I'm pretty sure there are missing commits too, since this change is adjusting two files that do not even exist now.

Any LGE device with a broadcom combo chip uses it.

Personally, my 2 cents are that commonizing it is a good idea; devices use different offsets, so if we have board configuration flags to set those, would be nice.

Obviously, the biggest part of this would be migration. Do we break existing installs? Continuity must be a thing.

Patch-set: 2