Update patch set 2
Patch Set 2:

This is the one where everything becomes configurable and therefore has the ifdefs to enable/disable.  Presently I'm planning to rename the executable to "hwaddrs2" to make enable/disable function properly.  I'm unsure I can get the setting of "ro.bt.bdaddr_path" to enable/disable properly and so that may need to remain with the devices (yuck, but if that is the only route).

Patch-set: 2
Reviewer: Gerrit User 18177 <18177@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>