Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> > Patch Set 2: Verified-1 Code-Review-2
> > 
> > Not really a good idea. This change is already in the msm8996 tree. Other devices that need it will do their own. I'm not even sure the hwaddrs is even applicable for all the LGE devices even. I'm pretty sure there are missing commits too, since this change is adjusting two files that do not even exist now.
> 
> Any LGE device with a broadcom combo chip uses it.
> 
> Personally, my 2 cents are that commonizing it is a good idea; devices use different offsets, so if we have board configuration flags to set those, would be nice.
> 
> Obviously, the biggest part of this would be migration. Do we break existing installs? Continuity must be a thing.

More-so the not a good idea was the entire hidden/private working.  There is no sepolicy at all in lge-common.  I can only assume (as I cannot see what the entire commit thead is), it appears this was taken from another branch, as the lineage-17.1 has this change in for msm8996 already, so why not start with what is working and not try to pull in another branch that as questionable?  While it might be able to be commonized across a large number of LG devices, I'm not sure I agree since it would require testing a larger number of devices for any change to ensure nothing breaks.  I have all 3 of the msm8996 devices to test with, but do not have the g2/g3 models.

Patch-set: 3
Reviewer: Gerrit User 14737 <14737@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>