Update patch set 3
Patch Set 3: Verified+1 Code-Review+1

Finally into recent times on msm8996-common.  Originally showed up on lineage-15.1.

This has had a fair bit of modification from what was in the other repositories.  The original had a hand conversion from hexadecimal to decimal, likely due to the author being unaware the "0x" notation *originated* with C.  Going to C's 0x prefix was squashed in.

Since this meant to be device-independent, I'm using a canonical example prefix of 0xDEADBE.

My reading of the specifications is bit 1 (note: bit 0 is least-significant) should be set if this is randomly generated.  As that is what would be in use for such case, set that in the random case example.
I think it would be a Good Idea for someone else to take a look at the specifications and say whether my interpretation is correct.  The original author thought it should be clear, so we differ here.

Patch-set: 3
Reviewer: Gerrit User 18177 <18177@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=+1
Label: Code-Review=+1