Update patch set 7
Patch Set 7:

Even though this is officially against Lineage 17.1, 16.0 and 17.1 are at the same point in Git.  I've successfully tried this on a Lineage 16.0 *user* build and it functions.  A further adjustment might be needed for 17.1, but this is highly functional.

Patch-set: 7
Reviewer: Gerrit User 18177 <18177@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>