Update patch set 8
Patch Set 8:

Alessandro Astone was trying this on the G2.  What was confirmed to work ended up having to remove this commit.  I would like to keep this commit in the chain, but if it works only for some devices and breaks on others this portion may have to remain device-specific.

Patch-set: 8
Reviewer: Gerrit User 18177 <18177@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>