Update patch set 6
Patch Set 6:

The reason this initial commit places `hwaddrs` into the update_engine domain is the update_engine domain had the key permission and was available on lineage-16.0.

There is potential `hwaddrs` could be in vendor_init for lineage-16.0 instead.  Thing is the call to switch to that needs to be made.  I'll take advice as this really is meant to target as far back as lineage-15.1.

Patch-set: 6
Reviewer: Gerrit User 18177 <18177@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>