Update patch set 1
Patch Set 1: Verified+1 Code-Review+1

Theory of having this one is to be an intermediate domain for the lineage-16.0 branch.  Perhaps 16.0 is the older branch, but these changes have been tested on 16.0 and will likely work with 15.1.

As of 16.0 vendor_init was available, vendor_misc_writer was not.  In fact 16.0 doesn't have many options at all.  This has been tested, and is known to work.

Question is simply: Where to place `hwaddrs.readmisc` for 16.0?  Answer that and it is time to officially be in lineage-16.0 and lineage-17.1.  I doubt this is going further.

Patch-set: 1
Reviewer: Gerrit User 18177 <18177@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=+1
Label: Code-Review=+1