Update patch set 7
Patch Set 7:

This seems excessive to me, but I'm also of the position that it's futile to guard against users shooting themselves in the foot if said users are granted root access. It's not really possible to protect against a user determined to shoot themselves in the foot.

Patch-set: 7
Reviewer: Gerrit User 2936 <2936@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>