Update patch set 15
Patch Set 15:

I believe #272046 is where CR+-2 should be used for voting for/against including this in lineage-15.1 should be.  The reason being is *this* is where merging the appropriate commits into lineage-15.1 happens.  All the other commits marked lineage-15.1 are actually shared among all branches present in lge/common.

An argument for merging this into lineage-15.1 is that is allows devices which got left on cm-14.1 to be updated incrementally instead of whole-hog.  Simply dumping all likely cm-14.1 -> lineage-17.1 changes into a repository is an approach, but the result may be rather inferior.

I note the G4 and V10 have been left at cm-14.1, so this is not an academic discussion.  This commit is simply my noting the option exists and keeping the door open for others.

Patch-set: 15
Reviewer: Gerrit User 18177 <18177@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>